Stock model price 칼럼 데이터  타입 변경2

+++ src/snowball/models.py
@@ -1,6 +1,6 @@
 from datetime import date as dt_date
 
-from sqlalchemy import Date, Numeric, PrimaryKeyConstraint, String
+from sqlalchemy import Date, Float, PrimaryKeyConstraint, String
 from sqlalchemy.orm import Mapped, mapped_column
 
 from src.database import Base
@@ -11,7 +11,7 @@ class Stock(Base):
 
     date: Mapped[dt_date] = mapped_column(Date, nullable=False)
     ticker: Mapped[str] = mapped_column(String(10), nullable=False)
-    price: Mapped[float] = mapped_column(Numeric(13, 4), nullable=False)
+    price: Mapped[float] = mapped_column(Float, nullable=False)
 
     __table_args__ = (PrimaryKeyConstraint("date", "ticker"),)
